+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Prevent de-serialising `undefined` if the default element's content is not
   null.
+- No longer serialise an empty array in the JSON serialisers, instead the
+  content can be removed for consistency with other tools.
 
 # 0.20.0